+++ NEWS.md
@@ -1,5 +1,19 @@
 # reticulate (development version)
 
+- The S3 classes for some (rarely encountered) Python objects have changed.
+  Only Python objects with non-standard `__module__` values are affected.
+  If a Python object’s parent class’s `__module__` attribute does not resolve to a string,
+  reticulate:
+    - Attempts to resolve it from the class's class, if it's a metaclass.
+    - If no string can be resolved, reticulate no longer implicitly prepends
+      'python.builtin.' as the class prefix, instead it uses just the `__name__`.
+  (See #1686 for more context)
+
+- Added support for Python 3.13. Note that Python 3.13 removed support
+  for `classmethod` descriptors, which may affect the S3 class of
+  some Python objects that use metaclass properties to resolve a class’s
+  `__module__` or `__name__` attribute. (#1686)
+
 - `py_is_null_xptr()` and `[[` now load delayed modules (#1688).
 
 - Fixed error when attempting to use a python venv created with `uv` (#1678)